pyproject.toml
[project]
name = "exam-line-bot-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.1",
    "line-bot-sdk>=3.17.1",
]
